Extract Snippet component/styles out from Profile page
