+++ frontend/src/Pages/Profile.jsx
@@ -1,18 +1,15 @@
 import React, { useEffect } from 'react';
-import { Row, Col, Button, Card, Dropdown } from 'react-bootstrap';
-import { ThreeDots } from 'react-bootstrap-icons';
-import DropdownToggle from 'react-bootstrap/esm/DropdownToggle';
+import { Row, Col, Button } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
-import { useSnippets } from '../hooks';
 import { actions as modalActions } from '../slices/modalSlice.js';
 import { actions as editorActions } from '../slices/editorSlice.js';
 import { fetchData } from '../slices/userSlice.js';
 import classes from './Profile.module.css';
+import Snippet from '../components/Snippet/Snippet.jsx';
 
 function Profile() {
   const { t } = useTranslation();
-  const snippetApi = useSnippets();
   const dispatch = useDispatch();
   const userInfo = useSelector((state) => state.user.userInfo);
   const snippets = useSelector((state) => state.snippets.snippets);
@@ -23,24 +20,6 @@ function Profile() {
     } catch {
       return t('profile.successfulLoading');
     }
-  };
-
-  const handleDeleteConfirmation = (id) => {
-    dispatch(
-      modalActions.openModal({
-        type: 'confirmDelete',
-        item: { id },
-      }),
-    );
-  };
-
-  const handleSnippetRename = (id, name, code) => {
-    dispatch(
-      modalActions.openModal({
-        type: 'renameRepl',
-        item: { id, name, code },
-      }),
-    );
   };
 
   const handleEditProfile = () => {
@@ -57,15 +36,6 @@ function Profile() {
       modalActions.openModal({
         type: 'changePassword',
         item: { id: userInfo.id },
-      }),
-    );
-  };
-
-  const handleSnippetShare = (item) => {
-    dispatch(
-      modalActions.openModal({
-        type: 'sharingRepl',
-        item,
       }),
     );
   };
@@ -161,83 +131,13 @@ function Profile() {
               </Row>
               <Row xs={1} md={2} className="g-4 my-1">
                 {snippets.map(({ id, slug, name, code }) => (
-                  <Col xs lg="3" key={id}>
-                    <Card style={{ border: 0 }}>
-                      <Card.Header
-                        className={`d-flex justify-content-between ${classes.snippetHeader}`}
-                      >
-                        <p className="m-0 p-2">{name}</p>
-                        <Dropdown
-                          className={`${classes.snippetTools}`}
-                          id="snippet"
-                        >
-                          <DropdownToggle
-                            data-bs-toggle="dropdown"
-                            aria-expanded="false"
-                            className={`flex-grow-0 btn btn-primary ${classes.dropdown}`}
-                          >
-                            <ThreeDots color="white" size="16" />
-                            <span className="visually-hidden">
-                              {t('profile.editSnippet')}
-                            </span>
-                          </DropdownToggle>
-                          <Dropdown.Menu className={`${classes.dropdownMenu}`}>
-                            <Dropdown.Item
-                              className={`${classes.dropdownItem}`}
-                              onClick={() =>
-                                handleSnippetRename(id, name, code)
-                              }
-                            >
-                              {t('profile.renameReplButton')}
-                            </Dropdown.Item>
-                            <Dropdown.Item
-                              className={`${classes.dropdownItem}`}
-                              onClick={() => handleDeleteConfirmation(id)}
-                            >
-                              {t('profile.deleteReplButton')}
-                            </Dropdown.Item>
-                          </Dropdown.Menu>
-                        </Dropdown>
-                      </Card.Header>
-                      <Card.Body className={`${classes.snippetBody}`}>
-                        <Card.Text>
-                          {/* TODO: add a screenshot for snippet */}
-                        </Card.Text>
-                        <div className="d-flex justify-content-center gap-2">
-                          <Button
-                            className={`btn-sm p-1 ${classes.button}`}
-                            variant="primary"
-                            href={snippetApi.genViewSnippetLink(
-                              userInfo.login,
-                              slug,
-                            )}
-                          >
-                            {t('profile.openReplButton')}
-                          </Button>
-                          <Button
-                            className={`btn-sm p-1 ${classes.button}`}
-                            variant="primary"
-                            onClick={() =>
-                              handleSnippetShare({
-                                name,
-                                id,
-                                link: snippetApi.genViewSnippetLink(
-                                  userInfo.login,
-                                  slug,
-                                ),
-                                embedLink: snippetApi.genEmbedSnippetLink(
-                                  userInfo.login,
-                                  slug,
-                                ),
-                              })
-                            }
-                          >
-                            {t('profile.shareReplButton')}
-                          </Button>
-                        </div>
-                      </Card.Body>
-                    </Card>
-                  </Col>
+                  <Snippet
+                    key={id}
+                    id={id}
+                    slug={slug}
+                    name={name}
+                    code={code}
+                  />
                 ))}
               </Row>
             </div>
