+++ frontend/src/components/Snippet/Snippet.module.css
@@ -1,0 +1,75 @@
+.snippetHeader {
+  background: linear-gradient(
+    135deg,
+    #5e3ce1 11.52%,
+    #524edc 31.69%,
+    #434ae4 55.94%,
+    #3a45e2 76.74%,
+    #3745e6 91.55%
+  );
+}
+
+.snippetTools {
+  display: flex;
+  flex-direction: row;
+  mag-height: 38px;
+  gap: 15%;
+}
+
+.dropdown {
+  background: transparent;
+  border: 0;
+  padding: 0;
+  color: #136ef6;
+  font-size: 0;
+
+  --bs-btn-hover-bg: transparent;
+  --bs-btn-hover-color: #7c7cf7;
+  --bs-btn-active-bg: transparent;
+}
+
+.dropdownMenu {
+  --bs-dropdown-min-width: 7rem;
+  background: #222e51;
+}
+
+.dropdownItem {
+  color: #136ef6;
+  font-weight: 500;
+  -webkit-background-clip: text;
+  background-size: 300% 200%;
+  background-position: 150%;
+  transition: background-position 675ms ease;
+  background-image: linear-gradient(
+    135deg,
+    #5f17f1 11.52%,
+    #524edc 31.69%,
+    #7c7cf7 55.94%,
+    #121ff1 76.74%,
+    #3745e6 91.55%
+  );
+}
+
+.dropdownItem:hover {
+  background-position: 0 100%;
+  color: transparent;
+}
+
+.dropdown:focus {
+  --bs-btn-hover-bg: transparent;
+  color: #7c7cf7;
+}
+
+.snippetBody {
+  background: #222e51;
+}
+
+.button {
+  color: #fff;
+  background: #136ef6;
+  border: 1px solid #3e496b;
+  border-radius: 5px;
+}
+.button:hover {
+  --bs-btn-hover-bg: #7c7cf7;
+}
